[geometry] Bugfix in document generator. Details:
The XML generated by Doxygen contains parameters twice: first shortly ("memberdef -> param"), below more in detail ("memberdef -> detaileddescription -> para -> parameterlist -> parameteritem"). However, the second one is not always in the parameter-order, apparently. At least this was the case for simplify.

This change fixes that. We now get it from param, get the details in a separate list and copy the details to the original list such that order is preserved.

[SVN r79528]