Try to fix some links in circular buffer.
Changes some links to use the location independent 'boost:' links.

Also changes the build settings so that 'boost:' is used as a prefix when
generating docbook. The reason is a bit obscure. When generating documentation
with an index, the docbook documentation has to first be generated separately,
so that the index doesn't get confused with other documents. But when this
happens the boost.root parameter from the circular buffer jamfile is used -
which was the wrong location. So I changed it to use the prefix 'boost:'
which will later get converted to the correct location when building the
combined documentation. Sorry if that doesn't make any sense, but hopefully it
should fix some broken links.

[SVN r86174]